Полная переработка бота

+++ src/main/resources/db/changelog/v1/v1.0__create_tablet_log.sql
@@ -1,6 +1,6 @@
 CREATE TABLE if not exists public.t_log (
 	id serial4 NOT NULL,
-	chat_id int4 NULL,
+	chat_id bigint NULL,
 	create_stamp timestamp DEFAULT CURRENT_TIMESTAMP NULL,
 	message text NULL
 );
